PayloadBuilder: Fix retrieving KVs with a custom config
Commit "audio-hal: Changes to support multiple custom configs" [1]
changes the custom_key format to support multiple keys separated by
a semicolon and added a semicolon to all usages. However, arpal-lx
was only updated to split the custom_key for the usage in
ressourcemanager_*.xml by commit "pal: Support for multiple custom
configs" [2] and missed to update the handling of custom keys for
usecases. This lead to custom keys being not found since they
expected an additional semicolon at the end of the name which is not
provided in the configs (and should obviously not be provided there).

[1]: https://github.com/LineageOS/android_hardware_qcom_audio-ar/commit/f74ac82eafde33c4130bd4f0d54b6ed38338607f
[2]: https://github.com/LineageOS/android_vendor_qcom_opensource_arpal-lx/commit/b116426312480d65c195acb2cd1e41bb0812dcc0

Change-Id: I4f9b955ae0091b50399c36b178a0099639719c97
Signed-off-by: Fiqri Ardyansyah <fiqri191002@gmail.com>